remove redundant escape function (use helper)
