+++ src/lib/helpers/html.ts
@@ -1,9 +1,12 @@
-// https://stackoverflow.com/a/6234804/315168
-export function escapeHtml(unsafe: string): string {
-	return unsafe
-		.replace(/&/g, '&amp;')
-		.replace(/</g, '&lt;')
-		.replace(/>/g, '&gt;')
-		.replace(/"/g, '&quot;')
-		.replace(/'/g, '&#039;');
+const entities = {
+	'&': '&amp;',
+	'<': '&lt;',
+	'>': '&gt;',
+	'"': '&quot;',
+	"'": '&apos;'
+};
+
+export function escapeHtml(unsafe: string) {
+	// @ts-ignore
+	return unsafe.replace(/[&<>"']/g, (char) => entities[char]);
 }
